clean up messes
