+++ docs/playground/1/01_examples/copy.mdx
@@ -1,5 +1,5 @@
 ---
-title: Link
+title: Copy to Clipboard
 ---
 
 import CopyToClipboard from '../../../../src/components/copy-to-clipboard';
