+++ src/components/back-button.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+export const BackButton = () => {
+  return (
+    <li className="ml-0 mb-3">
+      <Link to="/" className="d-block py-1 align-middle small text-dark">
+        <Icon
+          iconName={iconNames.ARROW_LEFT}
+          className="fill-black mt-n1 mr-1"
+          width="12"
+          height="12"
+        />
+        Back
+      </Link>
+    </li>
+  );
+};
