+++ src/components/index.js
@@ -1,3 +1,4 @@
+import BackButton from './back-button';
 import CardDecks from './card-decks';
 import DevOnly from './dev-only';
 import DevFrontmatter from './dev-frontmatter';
@@ -13,9 +14,11 @@ import SideNavigation from './side-navigation';
 import TableOfContents from './table-of-contents';
 import TextBalancer from './text-balancer';
 import TopBar from './top-bar';
+import TreeNode from './tree-node';
 import VersionDropdown from './version-dropdown';
 
 export {
+  BackButton,
   CardDecks,
   DevOnly,
   DevFrontmatter,
@@ -31,5 +34,6 @@ export {
   TableOfContents,
   TextBalancer,
   TopBar,
+  TreeNode,
   VersionDropdown,
 };
