+++ src/components/left-nav.js
@@ -3,38 +3,12 @@ import { Link } from 'gatsby';
 import Icon, { iconNames } from './icon/';
 import VersionDropdown from './version-dropdown';
 import { products } from '../constants/products';
-
-const baseUrl = (path, depth) => {
-  return path
-    .split('/')
-    .slice(0, depth)
-    .join('/');
-};
+import { filterAndSortLinks, getBaseUrl } from '../constants/utils';
+import { BackButton, TreeNode } from './';
 
 const productIcon = path => {
   const product = path.split('/')[1];
   return products[product] ? products[product].iconName : null;
-};
-
-const filterAndSort = (nodes, url) => {
-  return nodes
-    .map(node => ({
-      title: node.frontmatter.title || 'TITLE NEEDED',
-      navTitle: node.frontmatter.navTitle,
-      path: node.fields.path,
-      items: [],
-      itemObj: {},
-    }))
-    .filter(node => node.path.includes(url))
-    .sort((a, b) => {
-      if (a.path < b.path) {
-        return -1;
-      }
-      if (a.path > b.path) {
-        return 1;
-      }
-      return 0;
-    });
 };
 
 const makeTree = edges => {
@@ -69,32 +43,6 @@ const orderTree = (tree, order) => {
     }
   }
   return result;
-};
-
-const SubList = ({ children, collapsed }) => {
-  if (collapsed) {
-    return null;
-  } else {
-    return (
-      <ul className="ml-4 list-unstyled align-items-center">{children}</ul>
-    );
-  }
-};
-
-const Back = () => {
-  return (
-    <li className="ml-0 mb-3">
-      <Link to="/" className="d-block py-1 align-middle small text-dark">
-        <Icon
-          iconName={iconNames.ARROW_LEFT}
-          className="fill-black mt-n1 mr-1"
-          width="12"
-          height="12"
-        />
-        Back
-      </Link>
-    </li>
-  );
 };
 
 const SectionHeading = ({ newList, path }) => {
@@ -144,49 +92,14 @@ const SectionHeadingWithVersions = ({ newList, path, versionArray }) => {
   );
 };
 
-const TreeNode = ({ node, path }) => {
-  if (!node.path) {
-    return (
-      <li
-        className="mt-3 mb-2 font-weight-bold text-muted text-uppercase small"
-        key={node.path}
-      >
-        {node.navTitle ? node.navTitle : node.title}
-      </li>
-    );
-  }
-
-  return (
-    <li className="ml-0 align-items-center" key={node.path}>
-      <div className="d-flex align-items-center">
-        <Link
-          to={node.path}
-          className={`d-inline-block py-1 align-middle lh-12 ${
-            path === node.path ? 'active font-weight-bold text-dark' : ''
-          }`}
-        >
-          {node.navTitle ? node.navTitle : node.title}
-        </Link>
-      </div>
-      {node.items.length > 0 && (
-        <SubList collapsed={!path.includes(node.path)}>
-          {node.items.map(subNode => (
-            <TreeNode node={subNode} path={path} key={subNode.path} />
-          ))}
-        </SubList>
-      )}
-    </li>
-  );
-};
-
 const LeftNav = ({ navLinks, path, versionArray, navOrder = null }) => {
   const newList = versionArray
-    ? filterAndSort(navLinks, baseUrl(path, 3))
-    : filterAndSort(navLinks, baseUrl(path, 2));
+    ? filterAndSortLinks(navLinks, getBaseUrl(path, 3))
+    : filterAndSortLinks(navLinks, getBaseUrl(path, 2));
   const tree = orderTree(makeTree(newList), navOrder);
   return (
     <ul className="list-unstyled mt-0">
-      <Back />
+      <BackButton />
       {versionArray ? (
         <SectionHeadingWithVersions
           newList={newList}
